requirements.txt
dedalus-labs
python-dotenv
python-dateutil
icalendar

